build: Replace #!/usr/bin/env python with passed in PYTHON=
This means that if we were forced to use a specific python for the build, we
will put that binary into the top of samba-tool, so it continues to work
after the install.

Andrew Bartlett

Reviewed-by: Michael Adam <obnox@samba.org>
(cherry picked from commit 58e3c5323e343dcab1c528c6b6a44925b76cb297)

The last 2 patches address bug #9785 - Fixes to use specified python for runtime
installation of Samba.