build: Build all of samba in autoconf make test
The test system actually depends on far more than smbtorture these days
and this was masked by the build groups.  Rather than try and specify
everything that could be used, just build the lot (which is what was
essentially being done anyway).

This prepares for the removal of the libraries and binaries build
groups, to assist in improving waf performance for single-binary builds.

Andrew Bartlett

Reviewed-by: Jelmer Vernooij <jelmer@samba.org>